fix: fix numfmt with bullet case
