+++ packages/sheets-ui/src/controllers/editor/__tests__/end-edit.controller.spec.ts
@@ -16,12 +16,12 @@
 
 import type { ICellData, IDocumentData, Injector, Univer, Workbook } from '@univerjs/core';
 import type { IFunctionService } from '@univerjs/engine-formula';
-import { CellValueType, IConfigService, IContextService, IResourceLoaderService, LocaleService, LocaleType, Tools } from '@univerjs/core';
+import { CellValueType, IConfigService, IContextService, IResourceLoaderService, isRichText, LocaleService, LocaleType, Tools } from '@univerjs/core';
 import { LexerTreeBuilder } from '@univerjs/engine-formula';
 import { SpreadsheetSkeleton } from '@univerjs/engine-render';
 import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest';
 import { normalizeString } from '../../utils/char-tools';
-import { getCellDataByInput, isRichText } from '../editing.render-controller';
+import { getCellDataByInput } from '../editing.render-controller';
 import { createTestBed } from './create-test-bed';
 import { IMockFunctionService, MockFunctionService } from './mock-function.service';
